Add files via upload
This project was created thanks to the teaching in my module of Python in the University of Stirling in the United Kingdom. It represent the exploring, understanding, cleaning, analysis and visualisation of the data of a Company. I am really happy to publish this as my first project in this amazing language of programation.